Moving dsb normalization to common
