--- tests/mpx/analysis/normalization/__init__.py
@@ -1,5 +1,0 @@
-"""
-Tests for the normalization modules
-
-Copyright © 2024 Pixelgen Technologies AB.
-"""
